Merged PR 84443: #808855 - Adds unreferenced refund support
To keep the NotNull validation for the SettleTransaction and VoidTransaction requests I had to extract the relatedTransactionId field into every implementing class.

Now in the RefundTransactionRequest the client needs to provide one of  relatedTransactionId (which makes the refund referenced) or payment option with credit card information or payment option with UPO ID.

That's why I introduced one more validation with the AtLeastOneFieldPresent annotation. If you think the additional validation is unnecessary I will remove it.

I tested referenced and unreferenced refund with a credit card information.

Related work items: #808855